Feature/actualizar metodo registra accion

+++ app/Http/Controllers/CategoriaController.php
@@ -3,30 +3,36 @@
 namespace App\Http\Controllers;
 
 use App\Models\Categoria;
-use App\Models\Producto;
 use App\Models\Registro;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\File;
 use Illuminate\Validation\Rule;
 
 class CategoriaController extends Controller
 {
     public function index(Request $request)
     {
+        if (request()->user('admin')->cannot('viewAny', Categoria::class)) {
+            return view('admin.categorias.index'); // salir sin enviar datos
+        }
+
         $categorias = Categoria::all();
         return view('admin.categorias.index', compact('categorias'));
     }
 
     public function create()
     {
+        if (request()->user('admin')->cannot('create', Categoria::class)) { abort(403); }
+
         return view('admin.categorias.create');
     }
 
     public function store(Request $request)
     {
+        if (request()->user('admin')->cannot('create', Categoria::class)) { abort(403); }
+
         $request->validate([
             'nombre_categoria'      => ['required', 'max:250', 'unique:categorias'],
-            'descripcion_categoria' => [], 
+            'descripcion_categoria' => [],
         ]);
 
         $categoria = Categoria::create([
@@ -34,23 +40,27 @@ class CategoriaController extends Controller
             'descripcion_categoria' => request('descripcion_categoria'),
         ]);
 
-        Registro::registrar_accion($categoria, 'categorias', 3);
+        Registro::registrar_accion($categoria, 'Crea nueva categoría');
 
-        session()->flash('success_create', 'Categoría creada exitosamente!');
+        session()->flash('success', 'Categoría creada exitosamente!');
 
         return redirect('/admin/categorias/');
     }
 
     public function edit(Categoria $categoria)
     {
+        if (request()->user('admin')->cannot('update', Categoria::class)) { abort(403); }
+
         return view('admin.categorias.edit', compact('categoria'));
     }
 
     public function update(Request $request, Categoria $categoria)
     {
+        if (request()->user('admin')->cannot('update', Categoria::class)) { abort(403); }
+
         $request->validate([
             'nombre_categoria' => ['required', 'max:250', Rule::unique('categorias')->ignore($categoria->id)],
-            'descripcion_categoria' => [], 
+            'descripcion_categoria' => [],
         ]);
 
         $categoria->update([
@@ -58,16 +68,23 @@ class CategoriaController extends Controller
             'descripcion_categoria' => $request->input('descripcion_categoria'),
         ]);
 
-        Registro::registrar_accion($categoria, 'categorias', 4);
-        session()->flash('success_update', 'Categoría actualizada exitosamente!');
+        Registro::registrar_accion($categoria, 'Edita una categoría');
+
+        session()->flash('success', 'Categoría actualizada exitosamente!');
 
         return redirect()->route('categorias.index');
     }
 
     public function destroy(Categoria $categoria)
     {
+        if (request()->user('admin')->cannot('delete', Categoria::class)) { abort(403); }
+
         $categoria->delete();
-        session()->flash('success_delete', '¡Categoría eliminada exitosamente!');
+
+        Registro::registrar_accion($categoria, 'Elimina una categoría');
+
+        session()->flash('success', '¡Categoría eliminada exitosamente!');
+
         return redirect()->route('categorias.index');
     }
 }
